src/embeddings_pipeline/requirements.txt
qdrant-client
fastapi
uvicorn
python-dotenv
tenacity